requirements.txt
numpy==1.16.1
python-Levenshtein==0.12.0
scikit-learn==0.21.3
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tqdm==4.35.0
